+++ pkg/test/mocks/requests.go
@@ -1,17 +1,11 @@
 package mocks
 
 import (
-	"bytes"
-	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"encoding/hex"
-	"errors"
-	"fmt"
+	globalutils "github.com/4chain-ag/go-bsv-middleware/pkg/utils"
 
 	"github.com/4chain-ag/go-bsv-middleware/pkg/temporary/wallet"
 	"github.com/4chain-ag/go-bsv-middleware/pkg/transport"
-	"github.com/4chain-ag/go-bsv-middleware/pkg/transport/utils"
 )
 
 // Headers is a map of headers
@@ -71,111 +65,17 @@ func (rb *RequestBody) AuthMessage() *transport.AuthMessage {
 
 // PrepareInitialRequestBody prepares the initial request body
 func PrepareInitialRequestBody(mockedWallet wallet.WalletInterface) *RequestBody {
-	opts := wallet.GetPublicKeyOptions{IdentityKey: true}
-	clientIdentityKey, err := mockedWallet.GetPublicKey(context.Background(), opts)
-	if err != nil {
-		panic(err)
-	}
-
-	initialNonce, err := mockedWallet.CreateNonce(context.Background())
-	if err != nil {
-		panic(err)
-	}
-
-	initialRequest := transport.AuthMessage{
-		Version:      "0.1",
-		MessageType:  "initialRequest",
-		IdentityKey:  clientIdentityKey,
-		InitialNonce: initialNonce,
-	}
+	initialRequest := globalutils.PrepareInitialRequestBody(mockedWallet)
 
 	return NewRequestBody(initialRequest)
 }
 
 // PrepareGeneralRequestHeaders prepares the general request headers
 func PrepareGeneralRequestHeaders(mockedWallet wallet.WalletInterface, previousResponse *transport.AuthMessage, path, method string) (Headers, error) {
-	serverIdentityKey := previousResponse.IdentityKey
-	serverNonce := previousResponse.InitialNonce
-
-	opts := wallet.GetPublicKeyOptions{IdentityKey: true}
-	clientIdentityKey, err := mockedWallet.GetPublicKey(context.Background(), opts)
+	headers, err := globalutils.PrepareGeneralRequestHeaders(mockedWallet, previousResponse, path, method)
 	if err != nil {
-		return nil, errors.New("failed to get client identity key")
-	}
-
-	requestID := generateRandom()
-	encodedRequestID := base64.StdEncoding.EncodeToString(requestID)
-
-	newNonce, err := mockedWallet.CreateNonce(context.Background())
-	if err != nil {
-		return nil, errors.New("failed to create new nonce")
-	}
-
-	var writer bytes.Buffer
-
-	// Write the request ID
-	writer.Write(requestID)
-
-	// Write the method and path
-	err = utils.WriteVarIntNum(&writer, len(method))
-	if err != nil {
-		return nil, errors.New("failed to write method length")
-	}
-	writer.Write([]byte(method))
-
-	// Write the path
-	err = utils.WriteVarIntNum(&writer, len(path))
-	if err != nil {
-		return nil, errors.New("failed to write path length")
-	}
-	writer.Write([]byte(path))
-
-	// Write -1 (no query parameters)
-	err = utils.WriteVarIntNum(&writer, -1)
-	if err != nil {
-		return nil, errors.New("failed to write query parameters length")
-	}
-
-	// Write 0 (no headers)
-	err = utils.WriteVarIntNum(&writer, 0)
-	if err != nil {
-		return nil, errors.New("failed to write headers length")
-	}
-
-	// Write -1 (no body)
-	err = utils.WriteVarIntNum(&writer, -1)
-	if err != nil {
-		return nil, errors.New("failed to write body length")
-	}
-
-	signature, err := mockedWallet.CreateSignature(
-		context.Background(),
-		writer.Bytes(),
-		"auth message signature",
-		fmt.Sprintf("%s %s", newNonce, serverNonce),
-		serverIdentityKey,
-	)
-	if err != nil {
-		return nil, errors.New("failed to create signature")
-	}
-
-	headers := map[string]string{
-		"x-bsv-auth-version":      "0.1",
-		"x-bsv-auth-identity-key": clientIdentityKey,
-		"x-bsv-auth-nonce":        newNonce,
-		"x-bsv-auth-your-nonce":   serverNonce,
-		"x-bsv-auth-signature":    hex.EncodeToString(signature),
-		"x-bsv-auth-request-id":   encodedRequestID,
+		return nil, err
 	}
 
 	return headers, nil
-}
-
-func generateRandom() []byte {
-	b := make([]byte, 32)
-	_, err := rand.Read(b)
-	if err != nil {
-		panic(err)
-	}
-	return b
 }
